Trial implementation of path part highlight.
In order to tell follwing 2 entries presented by CtrlP.vim, I want to
highlight the different part differently.

  + src/main/java/com/example/www/materials/Iron.java
  + src/test/java/com/example/www/materials/Iron.java

The implementation assumes that the user specifies the name of directory
and the highlight applied to the name as follows in `.vimrc` file:

    let g:ctrlp_path_highlights = {
        \ "main": {"ctermfg": 27, "ctermbg": 118, "cterm": "bold"},
        \ "test": {"ctermbg": 127},
        \ }

Unfortunately, this implementation works so fine. The automatic detection of
similar paths sounds attractive, but not needed. The most challenging
part is not the detection of similarity. The consistency is challenging.
If we offered the automatic detection feature, then the user would offer
a sequence, not a set, of highlights. Then, `main` could be applied the
first hightlight at one point, the third highlight at another point
because the set of entries varies along to the growth of the code base.
How to consistently apply the first highlight to `main` is the
challenging issue.
If the user learned something like "hey, the directory name `main` is
ALWAYS shown by blue characters on a light green background on my
laptop!", then that is very effective. And the current approach can
easily offer that experience.

By the way, I used the shortest keywords because I came to know that
this is meaningful. The following topic is very impressive.

    https://zenn.dev/mattn/articles/a2f962733fad2f8613d1